@@ -28,7 +28,7 @@
     "copy-styles": "cpx \"src/**/*.scss\" dist/Styles"
   },
   "peerDependencies": {
-    "react": "^16.0.0"
+    "react": "^17.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.11.4",
@@ -47,8 +47,8 @@
     "@types/lodash.debounce": "^4.0.6",
     "@types/lodash.throttle": "^4.1.6",
     "@types/node": "^12.12.38",
-    "@types/react": "^16.9.27",
-    "@types/react-dom": "^16.9.7",
+    "@types/react": "^18.0.31",
+    "@types/react-dom": "^18.0.11",
     "@typescript-eslint/eslint-plugin": "^2.26.0",
     "@typescript-eslint/parser": "^2.26.0",
     "babel-eslint": "^10.0.3",
@@ -72,8 +72,8 @@
     "node-sass": "^4.14.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.4",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-is": "^16.13.1",
     "react-scripts": "^4.0.0",
     "sass-loader": "^10.0.1",
@@ -85,14 +85,13 @@
   ],
   "dependencies": {
     "@date-io/moment": "^1.3.13",
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/lab": "^4.0.0-alpha.61",
-    "@material-ui/pickers": "^3.3.10",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/material": "^5.11.15",
     "@types/google.maps": "^3.48.3",
     "@types/resize-observer-browser": "^0.1.5",
     "autosuggest-highlight": "^3.1.1",
     "lodash.debounce": "^4.0.8",
-    "material-table": "1.69.1",
     "moment": "^2.29.1",
     "react-number-format": "^4.6.3"
   }
